+++ src/lib/components/layout/Overlay/AccountPending.svelte
@@ -73,6 +73,7 @@
 						class="text-xs text-center w-full mt-2 text-gray-400 underline"
 						on:click={async () => {
 							localStorage.removeItem('token');
+							sessionStorage.removeItem('token');
 							location.href = '/auth';
 						}}>{$i18n.t('Sign Out')}</button
 					>