+++ src/routes/auth/+page.svelte
@@ -155,7 +155,7 @@
 
 	onMount(async () => {
 		const redirectPath = $page.url.searchParams.get('redirect');
-		if ($user !== undefined) {
+		if ($user) {
 			goto(redirectPath || '/');
 		} else {
 			if (redirectPath) {